+++ cocoa-beans-state/src/main/java/net/apartium/cocoabeans/state/numberic/IntObservable.java
@@ -1,0 +1,11 @@
+package net.apartium.cocoabeans.state.numberic;
+
+import net.apartium.cocoabeans.state.Observable;
+import org.jetbrains.annotations.ApiStatus;
+
+import java.util.function.IntSupplier;
+
+@ApiStatus.AvailableSince("0.0.47")
+public interface IntObservable extends Observable<Integer>, IntSupplier {
+
+}