+++ cocoa-beans-state/src/main/java/net/apartium/cocoabeans/state/numberic/LongObservable.java
@@ -1,0 +1,11 @@
+package net.apartium.cocoabeans.state.numberic;
+
+import net.apartium.cocoabeans.state.Observable;
+import org.jetbrains.annotations.ApiStatus;
+
+import java.util.function.LongSupplier;
+
+@ApiStatus.AvailableSince("0.0.47")
+public interface LongObservable extends Observable<Long>, LongSupplier {
+
+}